+++ src/griptape_nodes/common/node_executor.py
@@ -336,6 +336,7 @@ class NodeExecutor(EngineScoped):
                     parameter_values=dict(node.parameter_values),
                     node_metadata=cast("NodeMetadata", dict(node.metadata)),
                     variables=self._resolve_variables_for_node(node.name),
+                    local_object_source=node.local_object_source,
                     workflow_name=workflow_context.name,
                     workflow_file_path=workflow_context.file_path,
                     workflow_working_directory=workflow_context.working_directory,
@@ -2808,7 +2809,7 @@ class NodeExecutor(EngineScoped):
         if upstream_param.name in upstream_node.parameter_output_values:
             return upstream_node.parameter_output_values[upstream_param.name]
 
-        return upstream_node.get_parameter_value(upstream_param.name)
+        return upstream_node._get_raw_parameter_value(upstream_param.name)
 
     def _get_value_through_subflow_group_proxy(
         self,
@@ -2867,7 +2868,7 @@ class NodeExecutor(EngineScoped):
             if source_param.name in source_node.parameter_output_values:
                 value = source_node.parameter_output_values[source_param.name]
             else:
-                value = source_node.get_parameter_value(source_param.name)
+                value = source_node._get_raw_parameter_value(source_param.name)
 
             logger.debug(
                 "Traced through proxy: %s.%s -> %s.%s (value type: %s)",